Move confmap functionality to its internal package
